video from pictures/python_version/requirements.txt
# Core dependencies
PyQt6>=6.6.0
Pillow>=10.0.0
pydicom>=2.4.0
moviepy>=1.0.3

# Development and testing
pytest>=7.4.0
pip-audit>=2.6.0

# Optional for better performance
numpy>=1.24.0
